pyproject.toml
[project]
name = "bot01"
version = "0.1.0"
description = "bot01"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = ["nonebot-adapter-onebot>=2.4.6", "nonebot-adapter-console>=0.6.0", "nonebot2[fastapi]>=2.4.0"]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Console", module_name = "nonebot.adapters.console" }
]
plugins = []
plugin_dirs = ["bot01/plugins"]
builtin_plugins = ["echo"]
